Made auto-id concurrent/suspense safe
Will wait for a final solution from the React team but seriously, this is fine with me.

IDs are not set in the server render/hydration, but patched up in the client with useEffect. It’s unlikely for a user to interact with the element before the ID is applied.