Resubscribe node observer when nodeRef changes
Before this change, if you passed a new ref or changed the `ref.current` passed to `useRect` then you would never get rects from the new node. This update fixes this by resubscribing the listener if the node changes.

This is only a fix for the useRect hook, not the 2018-render-props thingy. It's incredible how easy reasoning with updates are with hooks vs render-props.

I'll take a look into fixing the `<Rect>` as well if you approve of this change.