[Combobox] Add proper ARIA labeling to docs and examples
"If the combobox has a visible label and the combobox element is an
HTML element that can be labelled using the HTML label element (e.g.,
the input element), it is labeled using the label element. Otherwise,
if it has a visible label, the combobox element has aria-labelledby
set to a value that refers to the labelling element. Otherwise, the
combobox element has a label provided by aria-label." [1]

Given the number of ways to label the combobox, we cannot enforce or
handle the aria programmatically. Outside of a runtime check, the best
we can do is set good examples with proper labeling.

[1]: https://www.w3.org/TR/wai-aria-practices-1.2/#wai-aria-roles-states-and-properties-6