[Combobox] Allow ARIA labels to go on parent Combobox
Labels for the compound component can go on the parent unless there's a
good reason to not provide this pattern for a specific component. One
benefit reaped from this pattern is that it alleviates the need for
developers to know about precisely which `aria-*` attributes belong on
which piece of a compound component—just pass all `aria-*` attributes to
the parent and let Reach handle putting it in the proper place.

Another benefit is that if the ARIA spec changes in the future (as it did from
1.1 to 1.2 for Combobox) Reach doesn't have to introduce a breaking API change.

Based on discussions had here: https://github.com/reach/reach-ui/pull/504