Remove unused use-throttle dep from yarn.lock
A fresh clone and install of this library results in a yarn.lock diff
with the `use-throttle` dep removed. This makes sense, as it's not used
anywhere in the source code nor is it a dep of any other dep.

The strange part is why this diff wasn't showing up sooner given that a
[blame][1] shows the most recent change to this part of the lock file
was 7 months ago.

[1]: https://github.com/reach/reach-ui/blame/881b6ee3496c4b5b868f5ca1f706af63ea48114e/yarn.lock#L16431