[menu-button] Fix preventing menu in dialog (#808)
Portaled menus still will not work correctly, but if this is needed
users can pass portal={false} to `MenuList` to prevent conflicts with
the dialog's focus lock. These changes make it easier to deal with
non-portaled popovers by allowing consumers to render the `Menu`
component as a DOM node that can be relatively positioned.

This change also removes the outdated `data-reach-menu` attribute from
the popover and uses it for the wrapping menu *if* a DOM node is
rendered. This is a breaking change.