fix: Make sure useId can be deterministic
If the user provides an ID fallback to a component that calls useId, we
need to make sure the IDs are deterministic beyond the first render. We
we mistakenly used the same ID across renders, so this is fixed. this
also adds a check for React's internal useId hook in React 18 and
prioritizes it over our hack. We don't formally support React 18 yet but
this will help get us there.