breaking(disclosure): Simplify state logic
In uncontrolled disclosures, previously we would always render the
disclosure as open on the first render, regardless of the defaultOpen
prop's value. This was viewed as a progressive enhancement to ensure
content was visible to users without JavaScript, but in more cases it
just meant that the user might see a flash before the state would
change, causing unwanted jank. We simplified this by simply respecting
the defaultOpen prop's value on the initial render. If you want the old
behavior, control the state of your disclosures and ensure that the
initial open state is always set to `true`, then set the desired state
for the following render in useEffect.