fix(tabs): Only manage focus w/ keyboard (#915)
Chrome will focus a button on click, but Safari will not. Becaus of this
we have previously managed the focus of keyboard interactions manually
for consistency between browsers. This breaks :focus-visible, as manual
.focus() calls trigger the focusring regardless of the input modality.
We have removed the manual focus call which makes tabs behave similar to
buttons in the same browser.